fix: derive unique encryption keys
Closes #5706

Using the same encryption key to encrypt an unlimited amount of
data/documents results in key wear-out, in particular when using AES-CBC
or other cipher modes that are not very resistant to nonce misuse.

This commit derives from the application secret a fresh encryption key
for every document to encrypt. HKDF is used to derive these keys and the
context info parameter is set to the fresh random IV value.

Using the IV value as the context info does not seem to violate the
technical requirements of HKDF which only requires that the info value
is not related to the IKM parameter, according to section 3.2 of
RFC5869. The context information is not required to be confidential
either and is in many cases a documented fixed value, so using a value
that is known by the attacker does not seem to be an issue. The only
thing that I can think of that  might be an issue is that the value is
not only known by the attacker, but the attacker can manipulate it.
However, modifying it would yield a completely different encryption key
and IV and I don't see what befenits they would gain from this. Famous
last words.

Using a random value as the context info instead of using it as the salt
parameter is a recommendation from
https://soatok.blog/2021/11/17/understanding-hkdf/